requirements.txt
gunicorn
openai==1.35.7
Flask-Cors==4.0.1
Flask==3.0.3
python-dotenv==1.0.1